fix race condition & simplify code
The mod previously added a tilesheet to the farm map as soon as it was available. Sometimes that happened when the map was still loading its tilesheets, which caused the load to crash with a collection-was-modified error. This commit ensures the patching only happens once the save is fully loaded.